+++ src/main/linear_intpl.cc
@@ -204,7 +204,7 @@ int main(int argc, char* argv[]) {
   }
 
   // Check x-axis values.
-  const int input_length(data_x.size());
+  const int input_length(static_cast<int>(data_x.size()));
   for (int i(1); i < input_length; ++i) {
     if (data_x[i] <= data_x[i - 1]) {
       std::ostringstream error_message;
